Rename: delete folder in components
