+++ src/pages/OrderCheck.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const OrderCheck = () => {
+  return <div>OrderCheck</div>;
+};
+
+export default OrderCheck;
